% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawl2-package.R
\docType{package}
\name{crawl2-package}
\alias{crawl2}
\alias{crawl2-package}
\title{crawl2: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Josh London \email{josh.london@noaa.gov}

}
\keyword{internal}
